requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
langchain-google-genai
jinja2
